Changed folder structure and combine reservation system elements
