+++ Cinecon/Misc/ConsoleHelper.cs
