+++ Cinecon/Misc/Models.cs
@@ -1,5 +1,6 @@
 ﻿using Newtonsoft.Json;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Cinecon
 {
@@ -44,6 +45,7 @@ namespace Cinecon
         public bool IsActivated { get; set; }
         [JsonProperty("payment_method")]
         public string PaymentMethod { get; set; }
+        public int Room { get; set; }
         public Movie Movie { get; set; }
         public List<Seat> Seats { get; set; }
     }
@@ -68,14 +70,14 @@ namespace Cinecon
         public List<Movie> Movies { get; set; }
         public List<Seat> Seats
         {
-            get 
+            get
             {
                 var seats = new List<Seat>();
                 for (int i = 0; i < TotalRows; i++)
                     for (int j = 0; j < SeatsPerRow; j++)
-                        seats.Add(new Seat { Row = ((char)(65 + i)).ToString(), Number = j + 1, IsTaken = false });
+                        seats.Add(new Seat { Row = ((char)(65 + i)).ToString(), Number = j + 1, IsTaken = JsonHelper.ReservationData?.Reservations?.Where(x => x.Room == Number).Any(x => x.Seats.Any(x => x.Row == ((char)(65 + i)).ToString() && x.Number == j + 1)) ?? false });
                 return seats;
-            } 
+            }
         }
     }
